fix bugs

+++ code/src/include/hash.h
@@ -1,9 +1,11 @@
 #ifndef BILSMTREE_HASH_H
 #define BILSMTREE_HASH_H
 
+#include <iostream>
+
 namespace bilsmtree {
 
-uint32_t Hash(const char* key, size_t len, uint32_t seed);
+uint Hash(const char* key, uint len, size_t seed);
 
 }
